workers/main.py
"""
LaraRadar Python Workers - FastAPI Main Application
Extended Threat Intelligence Platform Scanning Workers
"""

from fastapi import FastAPI, HTTPException, Depends, Header
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional, List, Dict
import os
from dotenv import load_dotenv
import redis
import httpx
from datetime import datetime

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="LaraRadar Workers API",
    description="Python scanning workers for Extended Threat Intelligence Platform",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure appropriately for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Redis connection
redis_client = redis.Redis(
    host=os.getenv("REDIS_HOST", "127.0.0.1"),
    port=int(os.getenv("REDIS_PORT", 6379)),
    db=int(os.getenv("REDIS_DB", 0)),
    decode_responses=True
)

# API Token validation
async def verify_token(x_api_token: str = Header(...)):
    """Verify API token from Laravel"""
    expected_token = os.getenv("LARAVEL_API_TOKEN")
    if x_api_token != expected_token:
        raise HTTPException(status_code=401, detail="Invalid API token")
    return x_api_token

# Pydantic models
class ScanRequest(BaseModel):
    domain: str
    scan_id: int
    options: Optional[Dict] = {}

class ScanResponse(BaseModel):
    success: bool
    message: str
    data: Optional[Dict] = None

class HealthResponse(BaseModel):
    status: str
    workers: Dict[str, str]
    version: str
    timestamp: str

# Health check endpoint
@app.get("/health", response_model=HealthResponse)
async def health_check():
    """Health check endpoint for worker status"""
    return {
        "status": "healthy",
        "workers": {
            "subdomain": "online",
            "port": "online",
            "tech": "online",
            "vuln": "online",
            "screenshot": "online",
            "darkweb": "online",
            "breach": "online"
        },
        "version": "1.0.0",
        "timestamp": datetime.now().isoformat()
    }

# Subdomain Enumeration Worker
@app.post("/workers/subdomain/scan", response_model=ScanResponse)
async def subdomain_scan(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Subdomain enumeration using Amass
    """
    try:
        # TODO: Implement Amass subdomain enumeration
        # For now, return mock data
        subdomains = [
            f"www.{request.domain}",
            f"mail.{request.domain}",
            f"ftp.{request.domain}",
            f"admin.{request.domain}",
        ]
        
        # Send results back to Laravel
        await send_to_laravel("/api/v1/subdomains/store", {
            "domain_id": request.scan_id,
            "subdomains": subdomains
        })
        
        return {
            "success": True,
            "message": f"Found {len(subdomains)} subdomains",
            "data": {"subdomains": subdomains}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Port Scanning Worker
@app.post("/workers/port/scan", response_model=ScanResponse)
async def port_scan(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Port scanning using Nmap
    """
    try:
        # TODO: Implement Nmap port scanning
        # Mock data for now
        ports = [
            {"port": 80, "service": "http", "state": "open"},
            {"port": 443, "service": "https", "state": "open"},
            {"port": 22, "service": "ssh", "state": "open"},
        ]
        
        return {
            "success": True,
            "message": f"Found {len(ports)} open ports",
            "data": {"ports": ports}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Technology Detection Worker
@app.post("/workers/tech/detect", response_model=ScanResponse)
async def tech_detect(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Technology detection using WhatWeb
    """
    try:
        # TODO: Implement WhatWeb technology detection
        technologies = [
            {"name": "Nginx", "version": "1.18.0", "category": "web-server"},
            {"name": "PHP", "version": "8.2", "category": "programming-language"},
        ]
        
        return {
            "success": True,
            "message": f"Detected {len(technologies)} technologies",
            "data": {"technologies": technologies}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Vulnerability Scanning Worker
@app.post("/workers/vuln/scan", response_model=ScanResponse)
async def vuln_scan(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Vulnerability scanning using Nuclei
    """
    try:
        # TODO: Implement Nuclei vulnerability scanning
        vulnerabilities = [
            {
                "title": "SSL Certificate Expired",
                "severity": "high",
                "cvss": 7.5,
                "description": "SSL certificate has expired"
            }
        ]
        
        return {
            "success": True,
            "message": f"Found {len(vulnerabilities)} vulnerabilities",
            "data": {"vulnerabilities": vulnerabilities}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Screenshot Worker
@app.post("/workers/screenshot/capture", response_model=ScanResponse)
async def screenshot_capture(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Capture screenshot using Selenium/Playwright
    """
    try:
        # TODO: Implement screenshot capture
        screenshot_path = f"/screenshots/{request.domain}.png"
        
        return {
            "success": True,
            "message": "Screenshot captured successfully",
            "data": {"screenshot_path": screenshot_path}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Dark Web Scraper Worker
@app.post("/workers/darkweb/scrape", response_model=ScanResponse)
async def darkweb_scrape(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Scrape dark web for credential leaks
    """
    try:
        # TODO: Implement dark web scraping
        credentials = []
        
        return {
            "success": True,
            "message": f"Found {len(credentials)} credential leaks",
            "data": {"credentials": credentials}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Breach Data Crawler Worker
@app.post("/workers/breach/check", response_model=ScanResponse)
async def breach_check(
    request: ScanRequest,
    token: str = Depends(verify_token)
):
    """
    Check for data breaches using LeakIX/HIBP APIs
    """
    try:
        # TODO: Implement LeakIX/HIBP API integration
        breaches = []
        
        return {
            "success": True,
            "message": f"Found {len(breaches)} breaches",
            "data": {"breaches": breaches}
        }
    except Exception as e:
        return {
            "success": False,
            "message": str(e),
            "data": None
        }

# Helper function to send data to Laravel
async def send_to_laravel(endpoint: str, data: dict):
    """Send scan results back to Laravel API"""
    laravel_url = os.getenv("LARAVEL_API_URL")
    api_token = os.getenv("LARAVEL_API_TOKEN")
    
    async with httpx.AsyncClient() as client:
        response = await client.post(
            f"{laravel_url}{endpoint}",
            json=data,
            headers={"Authorization": f"Bearer {api_token}"}
        )
        return response.json()

# Worker heartbeat (called periodically by scheduler)
@app.post("/workers/heartbeat")
async def worker_heartbeat(token: str = Depends(verify_token)):
    """Send heartbeat to Laravel to update worker status"""
    try:
        await send_to_laravel("/api/v1/workers/heartbeat", {
            "worker_name": "python-workers",
            "status": "online",
            "timestamp": datetime.now().isoformat()
        })
        return {"success": True, "message": "Heartbeat sent"}
    except Exception as e:
        return {"success": False, "message": str(e)}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8001)
