workers/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
redis==5.2.0
httpx==0.27.2
pydantic==2.10.0
python-multipart==0.0.12
celery==5.4.0
gunicorn==23.0.0

# Scanning tools Python wrappers
python-nmap==0.7.1
dnspython==2.7.0

# Screenshot tools
selenium==4.26.1
playwright==1.48.0

# Data processing
pandas==2.2.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
